Display 'Ad' icon for tabs opened via an ad or survey notification.
This uses the createTab callback (which can only be used for tabs asked to be created explicitly by browser-laptop, not from inside muon) to add a flag to the tabState about which tabs were created in response to an ad event.
This `isAdTab` property is reset when the tab receives a new Navigation event, i.e. the url is changed, either new origin or same origin.

Auditor:
@NejcZdovc